Refactor to use Phylo methods for searching the subtree
Phylo can be used directly to find clades whose name match with a given name using the `find_clades` method.
We only take the first match as the tree is supposed to not have any duplicate names.
We don't use indexes anymore as we can instead put the nodelist (later on) in a dictionary for fast lookup using clade names directly - this change might be reversed later, I'm just fiddling with the code right now to try and wrap my head around it.
All in all, this allow for a faster search of the subtree.