Refactor build nodelist
Refactor the nodelist construction so it takes less time.
Before this, for every node in the tree (2M items) we would:
- search once in the parasite list (50K items)
- search once in the freeliving list (40K items)
- open the csv file, write one row, close the csv file.

We now use two sets for constant time lookup of parasite/freeliving state of nodes. We write the complete nodelist at once at the end.

We did not keep compatibility with previous version in that indexes are not kept between the tree and the csv file, but reading the csv file into a dictionnary should alleviate any lookup time problem anyway.

Further work needs to be done though to actually implement this next part :)